requirements-visual.txt
streamlit>=1.28.0
Pillow>=9.0.0
numpy>=1.21.0,<2.0.0
scipy>=1.7.0
opencv-python-headless>=4.5.0
piexif>=1.1.3
scikit-image>=0.19.0
scikit-learn>=1.4.0
PyWavelets>=1.3.0
lightgbm>=3.3.0,<4.0.0
pandas>=1.3.0
pillow-heif>=0.13.0
joblib>=1.0.0
plotly>=5.0.0
